Change WPILib path to point to lib folder for Travis
When the environment variable TRAVIS is set to true, build.xml reads
lib/wpilib/java/current/ant/build.properties (in this repo). Before,
that file said wpilip was located under ${user.home}, which is not true
on Travis. Changed to say wpilib is located under the lib folder (in
this repo).